requirements.txt
git+https://github.com/aiogram/aiogram.git@dev-3.x
aiosqlite==0.19.0
asyncpg==0.30.0
bcrypt==4.1.1
python-dotenv==1.0.0
cloudinary==1.44.1
Pillow==11.0.0
flask==3.0.0
flask-cors==4.0.0

